feat: migrate catalog page
