+++ src/app/catalog/page.tsx
@@ -1,0 +1,11 @@
+import { Footer } from "../_components/_ui/Footer";
+import { Catalog } from "./_components/Catalog";
+
+export default function page() {
+  return (
+    <>
+      <Catalog />
+      <Footer className="bg-inherit" />
+    </>
+  );
+}
