--- src/components/Layouts/Catalog/type/index.ts
@@ -1,9 +1,0 @@
-export type TBanner = {
-  image: string;
-  title: string;
-  link: string;
-};
-
-export type TCatalog = TBanner & {
-  price: number;
-};
